move functions from ceph to ceph_manager
mount_osd_data and make_admin_daemon_dir are only used by
ceph_manager.py although they are defined in ceph.py

(cherry picked from commit f53ea258a4c86603f0982448b191488abb4e73f3)

Conflicts:
	tasks/ceph.py
